+++ CHANGELOG.md
@@ -11,9 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added the ability to switch from assertive (default) to polite aria modes, in the help menu (#309).
 - Added OpenAI GPT4-vision query system. Hit ? from the main chart to toggle on. (#317))
+- Added suggestions system for users to be able to more easily click
 
 ### Fixed
-
 
 - LLM popup now only triggered by ?, not /
 - LLM truncating responses, #322
@@ -26,7 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Commented out the instructions on GitHub templates so that users can keep it while adding new content (#308).
 - Added lineplot, stacked bar, dodged bar, and normalized dodge bar info to the README (#310).
 - Added Code of Conduct file in the project.
-
 
 ## [1.0.4] - 2023-11-30